requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
cbor==1.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
coloredlogs==15.0.1
comm==0.1.3
confection==0.1.0
contourpy==1.1.0
cwl-eval==1.0.12
cycler==0.11.0
cymem==2.0.7
Cython==0.29.36
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
faiss-gpu==1.7.2
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
ijson==3.2.2
importlib-metadata==6.8.0
importlib-resources==6.0.0
inscriptis==2.3.2
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
ir-datasets==0.5.5
ir-measures==0.3.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter-client==8.3.0
jupyter-core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
langcodes==3.3.0
lightgbm==3.3.5
lit==16.0.6
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nmslib==2.1.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyautocorpus==0.1.12
pybind11==2.6.1
pydantic==1.10.11
Pygments==2.15.1
pyjnius==1.5.0
pyparsing==3.0.9
pyserini==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrec-eval-terrier==0.5.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.12
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trec-car-tools==2.6
triton==2.0.0
typer==0.9.0
typing-extensions==4.7.1
tzdata==2023.3
unlzw3==0.2.2
urllib3==2.0.3
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wasabi==1.1.2
wcwidth==0.2.6
widgetsnbextension==4.0.8
yarl==1.9.2
zipp==3.16.0
zlib-state==0.1.5
